requirements.txt
numpy 
pandas
joblib
scikit-learn 
pyyaml
flask
